+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 ----------------
 
   * Update album name extraction.
+  * Update for Nuvola 4.9 API & Chromium.
 
 5.5 - September 28th, 2017
 -----------------------